sollidly-mvp/requirements.txt
customtkinter==5.2.1
openai==1.3.0
anthropic==0.7.0
py-hanspell
Pillow==10.1.0
python-dotenv==1.0.0
pynput==1.7.6
keyboard==0.13.5
pywin32==306
pyautogui==0.9.54
pystray==0.19.5
comtypes==1.2.0
